logika: fixed warnings

+++ logika/src/meters/logika4/4l/spt941_10.cpp
@@ -84,7 +84,7 @@ Spt941_10::Spt941_10(
 
 bool Spt941_10::IdentMatch( ByteType id0, ByteType id1, ByteType version ) const
 {
-    return Logika4L::IdentMatch( id0, id1, version ) && version < 0x80;
+    return Logika4L::IdentMatch( id0, id1, version ) && static_cast< uint8_t >( version ) < 0x80;
 } // IdentMatch
